\name{expr.tcga}
\docType{data}
\alias{expr.tcga}
\title{Tumor Bladder TCGA data}
\description{Counts matrix of RNA-seq study with 19 tumor Bladder
Urothelial Carcinoma samples and 19 adjacent healthy tissues}
\usage{expr.tcga}
\format{matrix}
\source{https://bioconductor.org/packages/release/bioc/html/GSEABenchmarkeR.html}
\references{Geistlinger L, Csaba G, Santarelli M, Ramos M, Schiffer L,
Law C, Turaga N,  Davis S, Carey V, Morgan M, Zimmer R, Waldron L
(2020). Toward a gold standard for benchmarking gene set enrichment
analysis. Briefings in Bioinformatics. doi:10.1093/bib/bbz158. }
\keyword{datasets}