\name{hallmarks.hs}
\docType{data}
\alias{hallmarks.hs}
\title{Hallmarks homo sapiens gene symbol}
\description{ Hallmark geneset collection from msigdb}
\usage{hallmarks.hs}
\value{
List with hallmark genes
}
\format{character list}
\source{https://www.gsea-msigdb.org/gsea/downloads.jsp}
\references{Liberzon, A. et al.: The Molecular Signatures
Database Hallmark Gene Set Collection. Cell Systems 1(6), 417-425 (2015). doi:10.1016/j.cels.2015.12.004}
\keyword{datasets}