\name{kegg.hs}
\docType{data}
\alias{kegg.hs}
\title{KEGG genesets homo sapiens entrez }
\description{KEGG genesets obtained with limma function getGeneKEGGLinks}
\usage{kegg.hs}
\value{
List with KEGG genes
}
\format{character list}
\source{https://www.kegg.jp/kegg/rest/keggapi.html}
\references{Kanehisa, M. and Goto, S.; KEGG: Kyoto Encyclopedia of Genes
and Genomes. Nucleic Acids Res. 28, 27-30 (2000).}
\keyword{datasets}