\name{pd.tcga}
\docType{data}
\alias{pd.tcga}
\title{Tumor Bladder TCGA data}
\description{Sample information of RNA-seq study with 19 tumor Bladder
Urothelial Carcinoma samples and 19 adjacent healthy tissues}
\usage{pd.tcga}
\value{
    Data frame with sample info
}
\format{DFrame}
\source{
https://bioconductor.org/packages/release/bioc/html/GSEABenchmarkeR.html
}
\references{Geistlinger L, Csaba G, Santarelli M, Ramos M, Schiffer L,
Law C, Turaga N,  Davis S, Carey V, Morgan M, Zimmer R, Waldron L
(2020). Toward a gold standard for benchmarking gene set enrichment
analysis. Briefings in Bioinformatics. doi:10.1093/bib/bbz158. }
\keyword{datasets}