
\name{roastgsa-internal}
\alias{avgcorfun}
\alias{contMatFunction}
\alias{designFunction}
\alias{dsfunction}
\alias{enrichmentScore}
\alias{formatApval}
\alias{froastgsa}
\alias{indexFunction}
\alias{make.contrasts}
\alias{sortDragHtmlTable.int}
\alias{write.html.mod}
\alias{heatmaprgsa}
\alias{expr.tcga}
\alias{fd.tcga}
\alias{pd.tcga}
\alias{kegg.hs}
\alias{sorttable}
\alias{dragtable}
\alias{hallmarks.hs}

\title{Internal roastgsa Functions}
\description{
Internal roastgsa functions
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\value{
Internal object
}
\keyword{ts}
\examples{
## do not run
numeric()
}