\name{sorttable}
\docType{data}
\alias{sorttable}
\title{sorttable for html writings }
\description{from sorttable v2.0 of  Stuart Langridge.}
\usage{sorttable}
\format{character vector}
\source{http://www.kryogenix.org/code/browser/sorttable/}
\references{http://www.kryogenix.org/code/browser/sorttable/}
\keyword{datasets}